Changed delegate back to old inheritance design
